Trait pallet_remote_asset_manager::pallet::Config

pub trait Config: Config + MaybeAssetIdConvert<u8, Self::AssetId> {

    type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + Into<u128>;
    type AssetId: Parameter + Member + Copy + MaybeSerializeDeserialize;
    type AssetIdConvert: Convert<Self::AssetId, Option<MultiLocation>>;
    type PalletStakingCallEncoder: StakingCallEncoder<<Self::Lookup as StaticLookup>::Source, Self::Balance, Self::AccountId, Context = Self::AssetId>;
    type PalletProxyCallEncoder: ProxyCallEncoder<Self::AccountId, ProxyType, Self::BlockNumber, Context = Self::AssetId>;
    type MinimumStatemintTransferAmount: Get<Self::Balance>;
    type SelfAssetId: Get<Self::AssetId>;
    type SelfLocation: Get<MultiLocation>;
    type SelfParaId: Get<ParaId>;
    type RelayChainAssetId: Get<Self::AssetId>;
    type AssetUnbondingSlashingSpans: Get<u32>;
    type AssetStakingCap: StakingCap<Self::AssetId, Self::Balance>;
    type Assets: MultiCurrency<Self::AccountId, CurrencyId = Self::AssetId, Balance = Self::Balance>;
    type XcmExecutor: ExecuteXcm<<Self as Config>::Call>;
    type XcmAssetTransfer: XcmTransfer<Self::AccountId, Self::Balance, Self::AssetId>;
    type AdminOrigin: EnsureOrigin<Self::Origin>;
    type XcmSender: SendXcm;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + Into<u128>

The balance type for cross chain transfers

type AssetId: Parameter + Member + Copy + MaybeSerializeDeserialize

Asset Id that is used to identify different kinds of assets.

type AssetIdConvert: Convert<Self::AssetId, Option<MultiLocation>>

Convert a T::AssetId to its relative MultiLocation identifier.

type PalletStakingCallEncoder: StakingCallEncoder<<Self::Lookup as StaticLookup>::Source, Self::Balance, Self::AccountId, Context = Self::AssetId>

The encoder to use for encoding when transacting a pallet_staking Call

type PalletProxyCallEncoder: ProxyCallEncoder<Self::AccountId, ProxyType, Self::BlockNumber, Context = Self::AssetId>

The encoder to use for encoding when transacting a pallet_proxy Call

type MinimumStatemintTransferAmount: Get<Self::Balance>

Minimum amount that can be transferred via XCM to the statemint parachain

type SelfAssetId: Get<Self::AssetId>

The native asset id

type SelfLocation: Get<MultiLocation>

The location of the chain itself

type SelfParaId: Get<ParaId>

Returns the parachain ID we are running with.

type RelayChainAssetId: Get<Self::AssetId>

Identifier for the relay chain’s specific asset

type AssetUnbondingSlashingSpans: Get<u32>

Unbonding slashing spans for unbonding on the relaychain.

type AssetStakingCap: StakingCap<Self::AssetId, Self::Balance>

Determines the threshold amounts when operating with staked assets.

type Assets: MultiCurrency<Self::AccountId, CurrencyId = Self::AssetId, Balance = Self::Balance>

Currency type for deposit/withdraw xcm assets

NOTE: it is assumed that the total issuance/total balance of an asset reflects the total balance of the PINT parachain account on the asset’s native chain

type XcmExecutor: ExecuteXcm<<Self as Config>::Call>

Executor for cross chain messages.

type XcmAssetTransfer: XcmTransfer<Self::AccountId, Self::Balance, Self::AssetId>

The type that handles all the cross chain asset transfers

type AdminOrigin: EnsureOrigin<Self::Origin>

Origin that is allowed to send cross chain messages on behalf of the PINT chain

type XcmSender: SendXcm

How to send an onward XCM message.

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type WeightInfo: WeightInfo

The weight for this pallet’s extrinsics.