Trait pallet_remote_asset_manager::traits::StakingCap

pub trait StakingCap<AssetId, Balance> {
    fn minimum_reserve_balance(asset: AssetId) -> Balance;

    fn minimum_bond_extra(asset: AssetId) -> Balance;
}

A type to abstract several staking related thresholds

Required methods

fn minimum_reserve_balance(asset: AssetId) -> Balance

The minimum amount that should be held in stash (must remain unbonded). Withdrawals are only authorized if the updated stash balance does exceeds this.

This must be at least the ExistentialDeposit as configured on the asset’s native chain (e.g. DOT/Polkadot)

fn minimum_bond_extra(asset: AssetId) -> Balance

The minimum required amount to justify an additional bond_extra XCM call to stake additional funds.